+++ vector2d.cpp
@@ -1,6 +1,6 @@
 #include "vector2d.h"
 
-Vector2D::Vector2D(double x, double y)
+Vector2D::Vector2D(float x, float y)
     :   x(x), y(y)
 {
 }
