+++ vector3d.cpp
@@ -1,6 +1,6 @@
 #include "vector3d.h"
 
-Vector3D::Vector3D(double x, double y, double z)
+Vector3D::Vector3D(float x, float y, float z)
     : x(x), y(y), z(z)
 {
 
@@ -13,7 +13,7 @@ Vector3D::Vector3D()
     z = 0;
 }
 
-double Vector3D::norm()
+float Vector3D::norm()
 {
     return qSqrt(x * x + y * y + z * z);
 }
@@ -22,16 +22,16 @@ Vector3D Vector3D::normalized() {
     return Vector3D(x / norm(), y / norm(), z / norm());
 }
 
-double Vector3D::dot_prod(Vector3D a, Vector3D b)
+float Vector3D::dot_prod(Vector3D a, Vector3D b)
 {
     return a.x * b.x + a.y * b.y + a.z * b.z;
 }
 
 Vector3D Vector3D::cross_prod(Vector3D a, Vector3D b)
 {
-    double x = a.y * b.z - b.y * a.z;
-    double y = -(a.x * b.z - b.x * a.z);
-    double z = a.x * b.y - b.x * a.y;
+    float x = a.y * b.z - b.y * a.z;
+    float y = -(a.x * b.z - b.x * a.z);
+    float z = a.x * b.y - b.x * a.y;
     return Vector3D(x, y, z);
 }
 
@@ -42,7 +42,7 @@ Vector3D operator+(const Vector3D & v1, const Vector3D v2) {
 Vector3D operator-(const Vector3D & v1, const Vector3D v2) {
     return v1 + -1*v2;
 }
-Vector3D operator*(const double & s, const Vector3D v) {
+Vector3D operator*(const float & s, const Vector3D v) {
     return Vector3D(s * v.x, s * v.y, s * v.z);
 }
